+++ nodalia-cards-2.1.1.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.1.2-alpha.3.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.2-alpha.3","contentSha256_12":"ba61578a77c1","file":"nodalia-cards-2.1.1.js","targetFile":"nodalia-cards-2.1.2-alpha.3.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.1.2-alpha.4.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.2-alpha.4","contentSha256_12":"7cb0527b2790","file":"nodalia-cards-2.1.1.js","targetFile":"nodalia-cards-2.1.2-alpha.4.js","fallbackFile":"nodalia-cards.js"};}